Update README.md (Add performance results from our Interspeech 2022 paper, SF-DST)
This commit adds the performance results from our paper presented at Interspeech 2022. The results are placed right below the SimpleTOD benchmark for comparison. A hyperlink to the paper is included, and the performance metrics are directly sourced from the publication.

Though some time has passed since its publication, I hope this contribution will still be valuable and find its place on the leaderboard.

This paper stands out as it incorporates the innovative use of in-context learning through questions and the self-feeding mechanism, which enhances the model’s ability to iteratively refine and improve its predictions. These elements contribute significantly to advancing the performance and understanding of task-oriented dialogue systems.